riscv: port kernel to 32-bit RISC-V (RV32)
- Sv32 page tables: 32-bit entries, PGD_SHIFT 22, PMD folded onto the
  PGD entry (PTRS_PER_PMD 1), 4MiB superpages; the early boot map uses
  4KiB pages since the kernel loads at 0x80200000 (not 4MiB aligned)
- 32-bit trap/switch frames (reg_t) and ld/sd -> lw/sw assembly
- xlen-aware CSR handling (unsigned long for XLEN-wide CSRs), satp
  (Sv32 mode 1 / Sv39 mode 8), TRAP_IS_INTERRUPT bit 31 vs 63
- RV32 VA layout: user 0-2G, direct map at 0xc0000000, KERNEL_LINK_ADDR
  0xc0000000; -mcmodel=medany for PC-relative pre-paging boot code
- Elf32 loader support and EI_CLASS validation
- type sweep: uintptr_t for pointer<->integer conversions, PRIx64/PRIu64
  format macros, phys_to_virt/page_to_virt return uintptr_t
- build: XLEN={32|64} selection, rv32 ld emulation, libgcc for 64-bit
  divide helpers, per-XLEN build dirs and rootfs images
- scripts: accept riscv32 toolchains, rv32 GDB arch, mkrootfs XLEN
- fix early page-table allocator reusing tables across map ranges
- sys_uname/arch_uname_machine and /proc/version report riscv32

+++ src/arch/riscv/include/arch/mm.h
@@ -21,39 +21,39 @@ extern char hart_entry[];
 
 static inline uint64_t symbol_phys(void *vaddr)
 {
-	return (uint64_t)vaddr - load_offset;
+	return (uintptr_t)vaddr - load_offset;
 }
 
 #define _SKERNEL_PHYS symbol_phys(_skernel)
 #define _EKERNEL_PHYS symbol_phys(_ekernel)
-#define _KERNEL_SIZE ((uint64_t)_ekernel - (uint64_t)_skernel)
+#define _KERNEL_SIZE ((uintptr_t)_ekernel - (uintptr_t)_skernel)
 
 #define _STEXT_PHYS symbol_phys(_stext)
 #define _ETEXT_PHYS symbol_phys(_etext)
 #define _ETEXT_ALIGNED_PHYS round_up(_ETEXT_PHYS, PAGE_SIZE)
-#define _TEXT_SIZE ((uint64_t)_etext - (uint64_t)_stext)
+#define _TEXT_SIZE ((uintptr_t)_etext - (uintptr_t)_stext)
 
 #define _SRODATA_PHYS symbol_phys(_srodata)
 #define _ERODATA_PHYS symbol_phys(_erodata)
 #define _ERODATA_ALIGNED_PHYS round_up(_ERODATA_PHYS, PAGE_SIZE)
-#define _RODATA_SIZE ((uint64_t)_erodata - (uint64_t)_srodata)
+#define _RODATA_SIZE ((uintptr_t)_erodata - (uintptr_t)_srodata)
 
 #define _SDATA_PHYS symbol_phys(_sdata)
 #define _EDATA_PHYS symbol_phys(_edata)
 #define _EDATA_ALIGNED_PHYS round_up(_EDATA_PHYS, PAGE_SIZE)
-#define _DATA_SIZE ((uint64_t)_edata - (uint64_t)_sdata)
+#define _DATA_SIZE ((uintptr_t)_edata - (uintptr_t)_sdata)
 
 #define _SBSS_PHYS symbol_phys(_sbss)
 #define _EBSS_PHYS symbol_phys(_ebss)
 #define _EBSS_ALIGNED_PHYS round_up(_EBSS_PHYS, PAGE_SIZE)
-#define _BSS_SIZE ((uint64_t)_ebss - (uint64_t)_sbss)
+#define _BSS_SIZE ((uintptr_t)_ebss - (uintptr_t)_sbss)
 
-static inline uint64_t phys_to_virt(uint64_t paddr)
+static inline uintptr_t phys_to_virt(uint64_t paddr)
 {
 	return (paddr - phys_ram_base) + PAGE_OFFSET;
 }
 
-static inline uint64_t virt_to_phys(uint64_t vaddr)
+static inline uint64_t virt_to_phys(uintptr_t vaddr)
 {
 	return (vaddr - PAGE_OFFSET) + phys_ram_base;
 }
